+++ packages/edit-site/src/components/sidebar-navigation-screen-patterns/category-item.js
@@ -14,15 +14,15 @@ export default function CategoryItem( {
 } ) {
 	const linkInfo = useLink(
 		{
-			path: '/library',
+			path: '/patterns',
 			categoryType: type,
 			categoryId: id,
 		},
 		{
 			// Keep a record of where we came from in state so we can
-			// use the browser's back button to go back to the library.
+			// use the browser's back button to go back to Patterns.
 			// See the implementation of the back button in patterns-list.
-			backPath: '/library',
+			backPath: '/patterns',
 		}
 	);