+++ packages/edit-site/src/components/sidebar-navigation-screen-patterns/index.js
@@ -4,13 +4,16 @@
 import {
 	__experimentalItemGroup as ItemGroup,
 	__experimentalItem as Item,
+	Flex,
+	Icon,
+	Tooltip,
 } from '@wordpress/components';
 import { useViewportMatch } from '@wordpress/compose';
 import { useSelect } from '@wordpress/data';
 import { getTemplatePartIcon } from '@wordpress/editor';
 import { __ } from '@wordpress/i18n';
 import { getQueryArgs } from '@wordpress/url';
-import { file, starFilled } from '@wordpress/icons';
+import { file, starFilled, lockSmall } from '@wordpress/icons';
 
 /**
  * Internal dependencies
@@ -19,7 +22,7 @@ import AddNewPattern from '../add-new-pattern';
 import SidebarNavigationItem from '../sidebar-navigation-item';
 import SidebarNavigationScreen from '../sidebar-navigation-screen';
 import CategoryItem from './category-item';
-import { DEFAULT_CATEGORY, DEFAULT_TYPE } from '../page-library/utils';
+import { DEFAULT_CATEGORY, DEFAULT_TYPE } from '../page-patterns/utils';
 import { store as editSiteStore } from '../../store';
 import { useLink } from '../routes/link';
 import usePatternCategories from './use-pattern-categories';
@@ -33,7 +36,7 @@ const templatePartAreaLabels = {
 	uncategorized: __( 'Uncategorized' ),
 };
 
-export default function SidebarNavigationScreenLibrary() {
+export default function SidebarNavigationScreenPatterns() {
 	const isMobileViewport = useViewportMatch( 'medium', '<' );
 	const { categoryType, categoryId } = getQueryArgs( window.location.href );
 	const currentCategory = categoryId || DEFAULT_CATEGORY;
@@ -68,7 +71,7 @@ export default function SidebarNavigationScreenLibrary() {
 	return (
 		<SidebarNavigationScreen
 			isRoot={ isTemplatePartsMode }
-			title={ __( 'Library' ) }
+			title={ __( 'Patterns' ) }
 			description={ __(
 				'Manage what patterns are available when editing your site.'
 			) }
@@ -76,11 +79,11 @@ export default function SidebarNavigationScreenLibrary() {
 			footer={ footer }
 			content={
 				<>
-					{ isLoading && __( 'Loading library' ) }
+					{ isLoading && __( 'Loading patterns' ) }
 					{ ! isLoading && (
 						<>
 							{ ! hasTemplateParts && ! hasPatterns && (
-								<ItemGroup className="edit-site-sidebar-navigation-screen-library__group">
+								<ItemGroup className="edit-site-sidebar-navigation-screen-patterns__group">
 									<Item>
 										{ __(
 											'No template parts or patterns found'
@@ -89,7 +92,7 @@ export default function SidebarNavigationScreenLibrary() {
 								</ItemGroup>
 							) }
 							{ hasMyPatterns && (
-								<ItemGroup className="edit-site-sidebar-navigation-screen-library__group">
+								<ItemGroup className="edit-site-sidebar-navigation-screen-patterns__group">
 									<CategoryItem
 										key={ myPatterns.name }
 										count={ myPatterns.count }
@@ -106,7 +109,7 @@ export default function SidebarNavigationScreenLibrary() {
 								</ItemGroup>
 							) }
 							{ hasTemplateParts && (
-								<ItemGroup className="edit-site-sidebar-navigation-screen-library__group">
+								<ItemGroup className="edit-site-sidebar-navigation-screen-patterns__group">
 									{ Object.entries( templatePartAreas ).map(
 										( [ area, parts ] ) => (
 											<CategoryItem
@@ -133,12 +136,38 @@ export default function SidebarNavigationScreenLibrary() {
 								</ItemGroup>
 							) }
 							{ hasPatterns && (
-								<ItemGroup className="edit-site-sidebar-navigation-screen-library__group">
+								<ItemGroup className="edit-site-sidebar-navigation-screen-patterns__group">
 									{ patternCategories.map( ( category ) => (
 										<CategoryItem
 											key={ category.name }
 											count={ category.count }
-											label={ category.label }
+											label={
+												<Flex
+													justify="left"
+													align="center"
+													gap={ 0 }
+												>
+													{ category.label }
+													<Tooltip
+														position="top center"
+														text={ __(
+															'Theme patterns cannot be edited.'
+														) }
+													>
+														<span className="edit-site-sidebar-navigation-screen-pattern__lock-icon">
+															<Icon
+																style={ {
+																	fill: 'currentcolor',
+																} }
+																icon={
+																	lockSmall
+																}
+																size={ 24 }
+															/>
+														</span>
+													</Tooltip>
+												</Flex>
+											}
 											icon={ file }
 											id={ category.name }
 											type="pattern"